requirements.txt
pyzmq
pandas
scikit-learn
pytest
mock